feat(user): implement get logged in user endpoint
This commit introduces a new endpoint `/me` to retrieve the
currently logged-in user's information. It uses the `verifyAuth`
middleware to ensure that only authenticated users can access
this endpoint. The `getLoggedInUser` controller function fetches
the user data based on the user ID stored in the context and
returns it in the response.
The console log was removed.